# docs/api/openapi.generated.json
{
  "components": {
    "schemas": {
      "APIErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "statusCode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statuscode"
          }
        },
        "required": [
          "code"
        ],
        "title": "APIErrorBody",
        "type": "object"
      },
...
      "OperationsSummary": {
        "additionalProperties": false,
        "properties": {
          "failedPaymentEvents": {
            "title": "Failedpaymentevents",
            "type": "integer"
          },
          "pendingPaymentEvents": {
            "title": "Pendingpaymentevents",
            "type": "integer"
          },
          "pendingWithdrawals": {
            "title": "Pendingwithdrawals",
            "type": "integer"
          },
          "processingWithdrawals": {
            "title": "Processingwithdrawals",
            "type": "integer"
          },
          "unverifiedAccounts": {
            "title": "Unverifiedaccounts",
            "type": "integer"
          }
        },
        "required": [
          "pendingWithdrawals",
          "processingWithdrawals",
          "failedPaymentEvents",
          "pendingPaymentEvents",
          "unverifiedAccounts"
        ],
        "title": "OperationsSummary",
        "type": "object"
      },
      "OrderBody": {
        "additionalProperties": false,
        "properties": {
          "customerId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Customerid",
            "type": "string"
          },
          "orderNo": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Orderno",
            "type": "string"
          },
          "profitAmount": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Profitamount"
          },
          "promoterId": {
            "format": "uuid",
            "title": "Promoterid",
            "type": "string"
          },
          "totalAmount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Totalamount"
          }
        },
        "required": [
          "orderNo",
          "customerId",
          "promoterId",
          "totalAmount",
          "profitAmount"
        ],
        "title": "OrderBody",
        "type": "object"
      },
      "OrderCreated": {
        "additionalProperties": false,
        "properties": {
          "orderId": {
            "title": "Orderid",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "paid",
              "failed",
              "refunded"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "orderId",
          "status"
        ],
        "title": "OrderCreated",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "与原 Fastify 路由和 JSON 契约兼容的 FastAPI 服务。",
    "title": "二级分销平台 API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/operations-summary": {
      "get": {
        "operationId": "operations_summary_api_admin_operations_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Not Implemented"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Operations Summary",
        "tags": [
          "financial lifecycle"
        ]
      }
    },
    "/api/admin/orders": {
      "post": {
        "operationId": "create_order_api_admin_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Not Implemented"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Order",
        "tags": [
          "financial lifecycle"
        ]
      }
    },
...
  }
}